Consolidated cursor handling into CursorUtil.cs
