+++ Assets/Scripts/Dialogue/DialogueManager.cs
@@ -242,7 +242,7 @@ public class DialogueManager : Manager<DialogueManager> {
     /// </summary>
     public static void Show(bool show) {
         Instance.gameObject.SetActive(show);
-        UIManager.ShowCustomCursor(!show);
+        CursorUtil.ShowCustomCursor(!show);
         UIManager.OnShowUIElement(show);
 
         if (!show) { Inventory.RevertSelection(); }
